requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
python-docx==1.1.2
python-multipart